+++ packages/eas-cli/src/credentials/ios/IosCredentialsProvider.ts
@@ -1,11 +1,16 @@
 import { Platform } from '@expo/eas-build-job';
 import { CredentialsSource, IosDistributionType, IosEnterpriseProvisioning } from '@expo/eas-json';
 
+import { CommonIosAppCredentialsFragment } from '../../graphql/generated';
+import Log from '../../log';
 import { findApplicationTarget } from '../../project/ios/target';
+import { confirmAsync } from '../../prompts';
 import { Context } from '../context';
 import * as credentialsJsonReader from '../credentialsJson/read';
 import { ensureAllTargetsAreConfigured } from '../credentialsJson/utils';
+import { getAppFromContext } from './actions/BuildCredentialsUtils';
 import { SetupBuildCredentials } from './actions/SetupBuildCredentials';
+import { SetupPushKey } from './actions/SetupPushKey';
 import { IosCapabilitiesOptions } from './appstore/ensureAppExists';
 import { App, IosCredentials, Target } from './types';
 import { isAdHocProfile } from './utils/provisioningProfile';
@@ -26,11 +31,14 @@ export default class IosCredentialsProvider {
   public async getCredentialsAsync(
     src: CredentialsSource.LOCAL | CredentialsSource.REMOTE
   ): Promise<IosCredentials> {
+    let buildCredentials;
     if (src === CredentialsSource.LOCAL) {
-      return await this.getLocalAsync();
+      buildCredentials = await this.getLocalAsync();
     } else {
-      return await this.getRemoteAsync();
+      buildCredentials = await this.getRemoteAsync();
     }
+    await this.getPushKeyAsync(this.ctx, this.options.targets);
+    return buildCredentials;
   }
 
   private async getLocalAsync(): Promise<IosCredentials> {
@@ -57,6 +65,40 @@ export default class IosCredentialsProvider {
       enterpriseProvisioning: this.options.enterpriseProvisioning,
       iosCapabilitiesOptions: this.options.iosCapabilitiesOptions,
     }).runAsync(this.ctx);
+  }
+
+  private async getPushKeyAsync(
+    ctx: Context,
+    targets: Target[]
+  ): Promise<CommonIosAppCredentialsFragment | null> {
+    if (ctx.nonInteractive) {
+      return null;
+    }
+
+    const applicationTarget = findApplicationTarget(targets);
+    const app = getAppFromContext(ctx);
+    const appLookupParams = {
+      ...app,
+      bundleIdentifier: applicationTarget.bundleIdentifier,
+      parentBundleIdentifier: applicationTarget.parentBundleIdentifier,
+    };
+
+    const setupPushKeyAction = new SetupPushKey(appLookupParams);
+    const isPushKeySetup = await setupPushKeyAction.isPushKeySetupAsync(ctx);
+    if (isPushKeySetup) {
+      Log.succeed(
+        `Push Notifications setup for ${app.projectName}:${applicationTarget.bundleIdentifier}`
+      );
+      return null;
+    }
+
+    const confirmSetup = await confirmAsync({
+      message: `Would you like to setup Push Notifications for your project?`,
+    });
+    if (!confirmSetup) {
+      return null;
+    }
+    return await setupPushKeyAction.runAsync(ctx);
   }
 
   private assertProvisioningProfileType(provisioningProfile: string, targetName?: string): void {
